fastapi/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
python-multipart
httpx